refactor: move github utils
